requirement.txt
Django==3.2.3
pillow==8.2.0
django-crispy-forms==1.11.2
django-tinymce==3.3.0
sslcommerz-python==0.0.7